refactor: Refactor encryption modules for Chromium browsers.
