+++ crypto/asn1pbe.go
@@ -6,8 +6,6 @@ import (
 	"crypto/sha256"
 	"encoding/asn1"
 	"errors"
-
-	"github.com/moond4rk/hackbrowserdata/utils/cryptoutil"
 )
 
 type ASN1PBE interface {
@@ -158,7 +156,7 @@ func (m metaPBE) deriveKeyAndIV(globalSalt []byte) ([]byte, []byte) {
 	iter := m.AlgoAttr.Data.Data.SlatAttr.IterationCount
 	keyLen := m.AlgoAttr.Data.Data.SlatAttr.KeySize
 
-	key := cryptoutil.PBKDF2Key(password[:], salt, iter, keyLen, sha256.New)
+	key := PBKDF2Key(password[:], salt, iter, keyLen, sha256.New)
 	iv := append([]byte{4, 14}, m.AlgoAttr.Data.IVData.IV...)
 	return key, iv
 }
